backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
google-genai>=0.12.0
python-dotenv>=1.0.1
pytest>=7.4.0
httpx>=0.24.1
prometheus-fastapi-instrumentator>=6.1.0
